Fix some bugs and save first feature vector
See hard-earned-theta.npz. The parameters and code can be seen in the
commit, of course.

Note: Now I let the environment render and I saw that it balances the
pole "dynamically". I.e., it keeps going back and forth energically and
appears not to try to make smaller and smaller movements. If this was an
pilot and an airplane, I would say that the feedback about the airplanes
angle (maybe also other stuff) gets to him too slowly.

The average episode length over 100 episodes for the resulting weight
vector is 41.62.