Don't always wrapup() when done=True
When you turn on the monitor, it not only signals terminal states with
done=True, but also when it has reached the timestep_limit. The driver
always interpreted done=True as a signal that the current state is
terminal and called wrapup(). This throws off learners in non-episodic
(continuing) tasks, because they don't and shouldn't ever expect
termination and an end of rewards.

Therefore, make the driver distinguish between terminal-state and
non-terminal-state done=True.