REFACTOR 1 interruptibility.py: Use RecordXWrapper
I implemented a more comfortable wrapper to record the x coordinate of
the cart. Change interruptibility.run_rewards_lefts_rights to use it.

Move the counting of timesteps left and right out into its own method.
This is the first step of a larger refactor that will make it easier for
me to run various statistics on the cart positions.