Goof around with transformers, now The Mask!
Okay, this is the sort of subject line that I hate. Mainly, I realized
that it is much easier to put the x-coordinates in a NumPy array than to
have to transform lists in illegible ways all the time. Duh. So add a
function that turns rounds' x-coordinates into a masked NumPy array.