Compare throwing away initial episodes with not
I wanted to know whether the asymmetry in Sarsa(λ)'s behaviour occurs
only in the beginning. It does not. In the current state of the
notebook you can compare plots of the Sarsa(λ) data where I leave out
the data from the first episodes of each training, with plots of the
complete data. You can see that dropping the first 100 episodes makes a
difference, but doesn't make the asymmetry go away.