linfa: Break ties randomly
I was wondering why my RL algorithm implementations behaved
asymmetrically. They tended to move the cart to the right. Finally I
found the cause, and it is quite simple: when all action values were the
same in some state, the algorithm chose one of them deterministically.
This made it asymmetric. Change it to choose randomly.